requirements.txt
requests_toolbelt==0.9.1
numpy==1.19.5
lxml==4.6.3
Brotli==1.0.9
cryptography==3.3.2
zipp==1.0.0
importlib_metadata==1.6.0
Django==2.2.19
tornado==6.1
beautifulsoup4==4.9.3
Cython==0.29.24
docutils==0.17.1
gtts==2.2.3
keyring==23.0.1
lockfile==0.12.2
monotonic==1.6
ordereddict==1.1
pdfminer==20191125
Pillow==8.3.1
protobuf==3.17.3
pyOpenSSL==20.0.1
pytesseract==0.3.8
pyttsx3==2.90
